refactored all templates to match inline styling
