+++ src/components/templates/footers/templateFooterTwo.jsx
@@ -1,111 +1,11 @@
 import React, { useState, useEffect } from "react";
-import { createUseStyles } from "react-jss";
 import AutoFillButton from "../autoFill/autoFillButton";
 import { useParams } from "react-router-dom";
 
-const useStyles = createUseStyles({
-  footer: {
-    border: "3px solid gray",
-    background:
-      "linear-gradient(45deg, #FCFCED 10%, #FCFCEE 15%, #161925 15%, #161925 100%)",
-    color: "#000",
-    fontFamily: "Montserrat, sans-serif",
-    fontWeight: "500",
-    padding: "1rem 2rem 1rem 2rem",
-    bottom: "-1",
-    left: "5vw",
-    minWidth: "90%",
-    width: "90%",
-    display: "flex",
-    marginBottom: "2rem",
-    justifyContent: "space-between",
-    alignItems: "center",
-    fontSize: "1vw",
-    "@media (max-width: 500px)": {
-      flexDirection: "column",
-      fontSize: "4vw",
-      fontFamily: "Arial, sans-serif",
-      fontWeight: "400",
-    },
-    "@media (max-width: 680px)": {
-      flexDirection: "column",
-      fontSize: "2vw",
-      fontFamily: "Arial, sans-serif",
-      fontWeight: "400",
-    },
-    "@media (max-width: 1500px)": {
-      background:
-        "linear-gradient(45deg, #161925 10%, #161925 25%, #FCFCEE 5%, #FCFCEE 100%)",
-    },
-    "@media (max-width: 800px)": {
-      background:
-        "linear-gradient(45deg, #161925 10%, #161925 29%, #FCFCEE 1%, #FCFCEE 100%)",
-    },
-    "@media (max-width: 750px)": {
-      background: "#FCFCEE",
-    },
-  },
-  leftContent: {
-    display: "flex",
-    alignItems: "center",
-    flexDirection: "column",
-    "& p": {
-      margin: "0 1rem",
-      "@media (max-width: 750px)": {
-        color: "#161925",
-      },
-    },
-    "& > a": {
-      textDecoration: "none",
-      color: "#FCFCEE",
-      "@media (max-width: 750px)": {
-        color: "#161925",
-      },
-    },
-    "@media (max-width: 500px)": {
-      paddingBottom: "1rem",
-    },
-  },
-  rightContent: {
-    display: "flex",
-    alignItems: "center",
-    flexDirection: "column",
-    color: "#fff",
-    "& p": {
-      margin: "0 1rem",
-    },
-    "& > a": {
-      textDecoration: "none",
-      color: "#161925",
-    },
-    "@media (max-width: 500px)": {
-      paddingBottom: "1rem",
-    },
-  },
-  middleContent: {
-    display: "flex",
-    color: "#fff",
-    alignItems: "center",
-    flexDirection: "column",
-    "& p": {
-      margin: "0 1rem",
-    },
-    "& > a": {
-      textDecoration: "none",
-      color: "#161925",
-    },
-    "@media (max-width: 500px)": {
-      paddingBottom: "1rem",
-    },
-  },
-});
-
 function TemplateFooterTwo() {
-  const classes = useStyles();
   const [isEditing, setEditing] = useState(false);
   const { templateId } = useParams();
 
- 
   const [details, setDetails] = useState({
     instagram: "Instagram",
     facebook: "Facebook",
@@ -173,8 +73,32 @@ function TemplateFooterTwo() {
   return (
     <>
       {showAutoFillButton && <AutoFillButton onAutoFill={handleAutoFill} />}
-      <footer className={classes.footer}>
-        <div className={classes.leftContent}>
+      <footer
+        style={{
+          border: "3px solid gray",
+          background: "linear-gradient(45deg, #FCFCED 10%, #FCFCEE 15%, #161925 15%, #161925 100%)",
+          color: "#000",
+          fontFamily: "Montserrat, sans-serif",
+          fontWeight: "500",
+          padding: "1rem 2rem 1rem 2rem",
+          bottom: "-1",
+          left: "5vw",
+          minWidth: "90%",
+          width: "90%",
+          display: "flex",
+          marginBottom: "2rem",
+          justifyContent: "space-between",
+          alignItems: "center",
+          fontSize: "1vw",
+        }}
+      >
+        <div
+          style={{
+            display: "flex",
+            alignItems: "center",
+            flexDirection: "column",
+          }}
+        >
           <p>Follow us:</p>
           {isEditing ? (
             <input
@@ -216,7 +140,14 @@ function TemplateFooterTwo() {
             <p>{leftContent.linkedIn}</p>
           )}
         </div>
-        <div className={classes.middleContent}>
+        <div
+          style={{
+            display: "flex",
+            alignItems: "center",
+            flexDirection: "column",
+            color: "#161925",
+          }}
+        >
           {isEditing ? (
             <input
               type="text"
@@ -244,7 +175,14 @@ function TemplateFooterTwo() {
             <p>{middleContent.contactEmail}</p>
           )}
         </div>
-        <div className={classes.rightContent}>
+        <div
+          style={{
+            display: "flex",
+            alignItems: "center",
+            flexDirection: "column",
+            color: "#fff",
+          }}
+        >
           <p>The Team:</p>
           {isEditing ? (
             <input
