+++ src/components/templates/headers/templateHeaderOne.jsx
@@ -3,190 +3,274 @@ import "../../../fonts.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { getAuth } from "firebase/auth";
 import { faBars } from "@fortawesome/free-solid-svg-icons";
-import { createUseStyles } from "react-jss";
 
 function TemplateHeaderOne() {
   // State variable to control the visibility of the responsive menu
   const [showMenu, setShowMenu] = useState(false);
-  const classes = useStyles();
+
   // Function to handle the click event on the bars icon
   const handleIconClick = () => {
     setShowMenu(!showMenu);
   };
+
   const auth = getAuth();
+
   const signOutUser = () => {
     auth.signOut();
   };
 
   return (
-    <nav className={classes.navParent}>
-      <h1 className={classes.h1}><span className={classes.craft}>Craft</span>My<span className={classes.portfolio}>Portfolio</span></h1>
-      <ul className={classes.menu}>
-        <li className={classes.MenuItem}>
-          <a href="/" className={classes.a}>
+    <nav
+      style={{
+        backgroundImage: `url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='1080' height='100' viewBox='0 0 1080 100'%3E%3Crect width='1080' height='100' fill='%23140021'%3E%3C/rect%3E%3Cpath d='M0 57L154 57L154 42L309 42L309 45L463 45L463 47L617 47L617 42L771 42L771 51L926 51L926 39L1080 39L1080 44L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%239900ff'%3E%3C/path%3E%3Cpath d='M0 63L154 63L154 56L309 56L309 63L463 63L463 53L617 53L617 52L771 52L771 61L926 61L926 66L1080 66L1080 58L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%238800e2'%3E%3C/path%3E%3Cpath d='M0 67L154 67L154 71L309 71L309 69L463 69L463 72L617 72L617 72L771 72L771 70L926 70L926 66L1080 66L1080 66L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%237700c6'%3E%3C/path%3E%3Cpath d='M0 82L154 82L154 82L309 82L309 80L463 80L463 79L617 79L617 74L771 74L771 79L926 79L926 82L1080 82L1080 76L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%236600ab'%3E%3C/path%3E%3Cpath d='M0 85L154 85L154 89L309 89L309 90L463 90L463 93L617 93L617 85L771 85L771 84L926 84L926 93L1080 93L1080 86L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%23560090'%3E%3C/path%3E%3C/svg%3E")`,
+        backgroundRepeat: "no-repeat",
+        backgroundSize: "cover",
+        display: "flex",
+        justifyContent: "space-between",
+        alignItems: "center",
+        padding: "1rem",
+      }}
+    >
+      <h1
+        style={{
+          fontSize: "48px",
+          fontWeight: 600,
+          color: "white",
+          fontFamily: "Delicious Handrawn, cursive",
+        }}
+      >
+        <span style={{ color: "white" }}>Craft</span>
+        My
+        <span style={{ color: "white" }}>Portfolio</span>
+      </h1>
+      <ul
+        style={{
+          display: "flex",
+          justifyContent: "space-between",
+          gap: "1rem",
+          listStyle: "none",
+          alignItems: "center",
+        }}
+      >
+        <li>
+          <a
+            href="/"
+            style={{
+              textDecoration: "none",
+              color: "white",
+              fontSize: "1.2rem",
+              fontWeight: 600,
+              transition: "all 0.1s ease-in-out",
+              borderRadius: "0.5rem",
+              cursor: "pointer",
+            }}
+          >
             Home
           </a>
         </li>
-        <li className={classes.MenuItem}>
-          <a href="/templates" className={classes.a}>
+        <li>
+          <a
+            href="/templates"
+            style={{
+              textDecoration: "none",
+              color: "white",
+              fontSize: "1.2rem",
+              fontWeight: 600,
+              transition: "all 0.1s ease-in-out",
+              borderRadius: "0.5rem",
+              cursor: "pointer",
+            }}
+          >
             Templates
           </a>
         </li>
-        <li className={classes.MenuItem}>
-          <a className={classes.a}>Account</a>
+        <li>
+          <a
+            style={{
+              textDecoration: "none",
+              color: "white",
+              fontSize: "1.2rem",
+              fontWeight: 600,
+              transition: "all 0.1s ease-in-out",
+              borderRadius: "0.5rem",
+              cursor: "pointer",
+            }}
+          >
+            Account
+          </a>
         </li>
-        <li className={classes.MenuItem}>
-          <a href="/details" className={classes.a}>
+        <li>
+          <a
+            href="/details"
+            style={{
+              textDecoration: "none",
+              color: "white",
+              fontSize: "1.2rem",
+              fontWeight: 600,
+              transition: "all 0.1s ease-in-out",
+              borderRadius: "0.5rem",
+              cursor: "pointer",
+            }}
+          >
             Details
           </a>
         </li>
-        <li className={`${classes.MenuItem} projects-button`}>
-          <a className={classes.a}>My Projects</a>
+        <li>
+          <a
+            className="projects-button"
+            style={{
+              textDecoration: "none",
+              color: "white",
+              fontSize: "1.2rem",
+              fontWeight: 600,
+              transition: "all 0.1s ease-in-out",
+              borderRadius: "0.5rem",
+              cursor: "pointer",
+            }}
+          >
+            My Projects
+          </a>
         </li>
-        <li className={classes.MenuItem} onClick={signOutUser}>
-          <a className={classes.a}>Log Out</a>
+        <li onClick={signOutUser}>
+          <a
+            style={{
+              textDecoration: "none",
+              color: "white",
+              fontSize: "1.2rem",
+              fontWeight: 600,
+              transition: "all 0.1s ease-in-out",
+              borderRadius: "0.5rem",
+              cursor: "pointer",
+            }}
+          >
+            Log Out
+          </a>
         </li>
       </ul>
       {/* Responsive menu for smaller screens */}
-      <div className={classes.resMenuParent}>
+      <div
+        style={{
+          display: "block",
+        }}
+      >
         <FontAwesomeIcon
-          className={classes.bars}
+          style={{
+            display: "none",
+          }}
           icon={faBars}
           onClick={handleIconClick}
         />
-        <ul className={showMenu ? classes.resMenu : classes.resMenuHide}>
-          <li className={classes.MenuItem}>
-            <a href="/" className={classes.a}>
+        <ul
+          style={{
+            display: "none",
+          }}
+        >
+          <li>
+            <a
+              href="/"
+              style={{
+                textDecoration: "none",
+                color: "white",
+                fontSize: "1.2rem",
+                fontWeight: 600,
+                borderRadius: "0.5rem",
+                cursor: "pointer",
+              }}
+            >
               Home
             </a>
           </li>
-          <li className={classes.MenuItem}>
-            <a href="/templates" className={classes.a}>
+          <li>
+            <a
+              href="/templates"
+              style={{
+                textDecoration: "none",
+                color: "white",
+                fontSize: "1.2rem",
+                fontWeight: 600,
+                borderRadius: "0.5rem",
+                cursor: "pointer",
+              }}
+            >
               Templates
             </a>
           </li>
-          <li className={classes.MenuItem}>
-            <a href="/details" className={classes.a}>
+          <li>
+            <a
+              href="/details"
+              style={{
+                textDecoration: "none",
+                color: "white",
+                fontSize: "1.2rem",
+                fontWeight: 600,
+                borderRadius: "0.5rem",
+                cursor: "pointer",
+              }}
+            >
               Details
             </a>
           </li>
-          <li className={classes.MenuItem}>
-            <a className={classes.a}>Account</a>
+          <li>
+            <a
+              style={{
+                textDecoration: "none",
+                color: "white",
+                fontSize: "1.2rem",
+                fontWeight: 600,
+                borderRadius: "0.5rem",
+                cursor: "pointer",
+              }}
+            >
+              Account
+            </a>
           </li>
-          <li className={`${classes.MenuItem} projects-button`}>
-            <a href="/details" className={classes.a}>
+          <li>
+            <a
+              style={{
+                textDecoration: "none",
+                color: "white",
+                fontSize: "1.2rem",
+                fontWeight: 600,
+                borderRadius: "0.5rem",
+                cursor: "pointer",
+              }}
+            >
               Details
             </a>
           </li>
-          <li className={`${classes.MenuItem} projects-button`}>
-            <a className={classes.a}>My Projects</a>
+          <li>
+            <a
+              className="projects-button"
+              style={{
+                textDecoration: "none",
+                color: "white",
+                fontSize: "1.2rem",
+                fontWeight: 600,
+                borderRadius: "0.5rem",
+                cursor: "pointer",
+              }}
+            >
+              My Projects
+            </a>
           </li>
-          <li className={classes.MenuItem} onClick={signOutUser}>
-            <a className={classes.a}>Log Out</a>
+          <li onClick={signOutUser}>
+            <a
+              style={{
+                textDecoration: "none",
+                color: "white",
+                fontSize: "1.2rem",
+                fontWeight: 600,
+                borderRadius: "0.5rem",
+                cursor: "pointer",
+              }}
+            >
+              Log Out
+            </a>
           </li>
         </ul>
       </div>
     </nav>
   );
 }
-
-const useStyles = createUseStyles({
-  navParent: {
-    backgroundImage: `url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='1080' height='100' viewBox='0 0 1080 100'%3E%3Crect width='1080' height='100' fill='%23140021'%3E%3C/rect%3E%3Cpath d='M0 57L154 57L154 42L309 42L309 45L463 45L463 47L617 47L617 42L771 42L771 51L926 51L926 39L1080 39L1080 44L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%239900ff'%3E%3C/path%3E%3Cpath d='M0 63L154 63L154 56L309 56L309 63L463 63L463 53L617 53L617 52L771 52L771 61L926 61L926 66L1080 66L1080 58L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%238800e2'%3E%3C/path%3E%3Cpath d='M0 67L154 67L154 71L309 71L309 69L463 69L463 72L617 72L617 72L771 72L771 70L926 70L926 66L1080 66L1080 66L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%237700c6'%3E%3C/path%3E%3Cpath d='M0 82L154 82L154 82L309 82L309 80L463 80L463 79L617 79L617 74L771 74L771 79L926 79L926 82L1080 82L1080 76L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%236600ab'%3E%3C/path%3E%3Cpath d='M0 85L154 85L154 89L309 89L309 90L463 90L463 93L617 93L617 85L771 85L771 84L926 84L926 93L1080 93L1080 86L1080 101L1080 101L926 101L926 101L771 101L771 101L617 101L617 101L463 101L463 101L309 101L309 101L154 101L154 101L0 101Z' fill='%23560090'%3E%3C/path%3E%3C/svg%3E")`,
-    backgroundRepeat: "no-repeat",
-    backgroundSize: "cover",
-    display: "flex",
-    justifyContent: "space-between",
-    alignItems: "center",
-    padding: "1rem",
-  },
-  h1: {
-    fontSize: "48px !important",
-    fontWeight: 600,
-    color: "white",
-    fontFamily: "Delicious Handrawn, cursive",
-  },
-  ul: {
-    display: "flex",
-    justifyContent: "space-between",
-    gap: "1rem",
-    listStyle: "none",
-    alignItems: "center",
-    "& li": {
-      "& a": {
-        fontFamily: "Delicious Handrawn, cursive !important",
-      },
-    },
-  },
-  menu: {
-    display: "flex",
-    justifyContent: "space-between",
-    gap: "1rem",
-    listStyle: "none",
-    alignItems: "center",
-    "& li": {
-      "& a": {
-        fontFamily: "Raleway, sans-serif !important",
-        "&:hover": {
-          textDecoration: "underline !important",
-          transition: "all 0.3s ease-in-out",
-        },
-      },
-    },
-  },
-  resMenuParent: {
-    display: "none",
-  },
-  resMenu: {
-    display: "none",
-  },
-  resMenuHide: {
-    display: "none",
-  },
-  a: {
-    textDecoration: "none",
-    color: "white",
-    fontSize: "1.2rem",
-    fontWeight: 600,
-    transition: "all 0.1s ease-in-out",
-    borderRadius: "0.5rem",
-    cursor: "pointer",
-    "&:hover": {
-      backgroundColor: "var(--white)",
-      color: "var(--accent-colour)",
-    },
-  },
-  bars: {
-    display: "none",
-  },
-  "@media screen and (orientation: portrait)": {
-    h1: {
-      fontSize: "20px !important",
-      color: "var(--text-colour)",
-    },
-    menu: {
-      display: "none",
-    },
-    resMenuParent: {
-      display: "block",
-    },
-    resMenu: {
-      display: "block",
-      position: "absolute",
-      padding: "1rem",
-      top: "3rem",
-      right: "1rem",
-      backgroundColor: "var(--primary-colour)",
-      borderRadius: "5px",
-      border: "2px solid var(--text-colour)",
-    },
-    bars: {
-      display: "block",
-      color: "var(--text-colour)",
-      fontSize: "2rem",
-      cursor: "pointer",
-    },
-  },
-});
 
 export default TemplateHeaderOne;
